flask/requirements.txt
cffi==1.14.1
click==7.1.2
cryptography==3.0
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
picamera==1.13
pycparser==2.20
pymysql==0.10.0
pytz==2020.1
six==1.15.0
sqlalchemy==1.3.18
werkzeug==1.0.1
wtforms==2.3.1
RPi.GPIO==0.7.0